refactor: 기존 themeProvider 이동
