+++ client/components/Logo/index.tsx
@@ -1,6 +1,6 @@
 import { css } from '@emotion/react';
 import { useRouter } from 'next/router';
-import { useGlobalTheme } from '../../styles/GlobalThemeContext';
+import { useGlobalTheme } from '../../styles/GlobalThemeContextLegacy';
 import LogoSvg from '/assets/logo.svg';
 
 interface LogoProps {
